refactor auth to separate module
